app/requirements.txt
langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
pandas>=2.2
openpyxl>=3.1
qdrant-client>=1.9
python-dotenv>=1.0
